QUẢN LÝ BÃI ĐỖ XE THÔNG MINH/parking_system/requirements.txt
Flask==2.3.3
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
torch==2.2.0
torchvision==0.17.0
easyocr==1.7.1 